+++ spotbugs/src/main/java/edu/umd/cs/findbugs/detect/RuntimeExceptionCapture.java
@@ -21,6 +21,7 @@
 package edu.umd.cs.findbugs.detect;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.BitSet;
 import java.util.Collection;
 import java.util.HashSet;
@@ -30,10 +31,13 @@ import java.util.Set;
 import org.apache.bcel.Const;
 import org.apache.bcel.classfile.Code;
 import org.apache.bcel.classfile.CodeException;
+import org.apache.bcel.classfile.ConstantPool;
 import org.apache.bcel.classfile.JavaClass;
+import org.apache.bcel.classfile.LineNumberTable;
 import org.apache.bcel.classfile.Method;
 import org.apache.bcel.generic.ASTORE;
 import org.apache.bcel.generic.InstructionHandle;
+import org.apache.bcel.generic.InstructionList;
 
 import edu.umd.cs.findbugs.BugAccumulator;
 import edu.umd.cs.findbugs.BugInstance;
@@ -46,6 +50,7 @@ import edu.umd.cs.findbugs.ba.BasicBlock;
 import edu.umd.cs.findbugs.ba.CFG;
 import edu.umd.cs.findbugs.ba.CFGBuilderException;
 import edu.umd.cs.findbugs.ba.DataflowAnalysisException;
+import edu.umd.cs.findbugs.ba.EdgeTypes;
 import edu.umd.cs.findbugs.ba.Hierarchy2;
 import edu.umd.cs.findbugs.ba.LiveLocalStoreDataflow;
 import edu.umd.cs.findbugs.ba.Location;
@@ -60,6 +65,8 @@ import edu.umd.cs.findbugs.classfile.Global;
 import edu.umd.cs.findbugs.classfile.MissingClassException;
 import edu.umd.cs.findbugs.internalAnnotations.DottedClassName;
 import edu.umd.cs.findbugs.util.ClassName;
+
+import static edu.umd.cs.findbugs.util.Util.isGeneratedCodeInCatchBlockViaLineNumber;
 
 /**
  * RuntimeExceptionCapture
@@ -87,6 +94,8 @@ public class RuntimeExceptionCapture extends OpcodeStackDetector implements Stat
         public boolean seen = false;
 
         public boolean dead = false;
+
+        public boolean rethrown = false;
 
         public ExceptionCaught(String exceptionClass, int startOffset, int endOffset, int sourcePC) {
             this.exceptionClass = exceptionClass;
@@ -121,6 +130,10 @@ public class RuntimeExceptionCapture extends OpcodeStackDetector implements Stat
     @Override
     public void visitAfter(Code obj) {
         for (ExceptionCaught caughtException : catchList) {
+            if (caughtException.rethrown) {
+                continue;
+            }
+
             Set<String> thrownSet = new HashSet<>();
             for (ExceptionThrown thrownException : throwList) {
                 if (thrownException.offset >= caughtException.startOffset && thrownException.offset < caughtException.endOffset) {
@@ -130,37 +143,40 @@ public class RuntimeExceptionCapture extends OpcodeStackDetector implements Stat
                     }
                 }
             }
-            int catchClauses = 0;
+
+            int priority = caughtException.dead ? NORMAL_PRIORITY : LOW_PRIORITY;
             if ("java.lang.Exception".equals(caughtException.exceptionClass) && !caughtException.seen) {
                 // Now we have a case where Exception is caught, but not thrown
                 boolean rteCaught = false;
                 for (ExceptionCaught otherException : catchList) {
                     if (otherException.startOffset == caughtException.startOffset
                             && otherException.endOffset == caughtException.endOffset) {
-                        catchClauses++;
                         if ("java.lang.RuntimeException".equals(otherException.exceptionClass)) {
                             rteCaught = true;
                         }
                     }
                 }
-                int range = caughtException.endOffset - caughtException.startOffset;
+
                 if (!rteCaught) {
-                    int priority = LOW_PRIORITY + 1;
-                    if (range > 300) {
-                        priority--;
-                    } else if (range < 30) {
-                        priority++;
-                    }
-                    if (catchClauses > 1) {
-                        priority++;
-                    }
-                    if (thrownSet.size() > 1) {
-                        priority--;
-                    }
-                    if (caughtException.dead) {
-                        priority--;
-                    }
                     accumulator.accumulateBug(new BugInstance(this, "REC_CATCH_EXCEPTION", priority).addClassAndMethod(this),
+                            SourceLineAnnotation.fromVisitedInstruction(getClassContext(), this, caughtException.sourcePC));
+                } else {
+                    accumulator.accumulateBug(new BugInstance(this, "REC2_CATCH_EXCEPTION", LOW_PRIORITY).addClassAndMethod(this),
+                            SourceLineAnnotation.fromVisitedInstruction(getClassContext(), this, caughtException.sourcePC));
+                }
+            } else if ("java.lang.RuntimeException".equals(caughtException.exceptionClass) && !caughtException.seen) {
+                accumulator.accumulateBug(new BugInstance(this, "REC2_CATCH_RUNTIME_EXCEPTION", LOW_PRIORITY).addClassAndMethod(this),
+                        SourceLineAnnotation.fromVisitedInstruction(getClassContext(), this, caughtException.sourcePC));
+            } else if ("java.lang.Throwable".equals(caughtException.exceptionClass) && !caughtException.seen) {
+                Method method = getMethod();
+                ConstantPool cp = method.getConstantPool();
+                LineNumberTable lineNumberTable = method.getLineNumberTable();
+                int line = lineNumberTable.getSourceLine(caughtException.sourcePC);
+                InstructionList list = new InstructionList(method.getCode().getCode());
+
+                if (!method.isSynthetic() &&
+                        !isGeneratedCodeInCatchBlockViaLineNumber(cp, lineNumberTable, line, list, Arrays.asList(obj.getExceptionTable()))) {
+                    accumulator.accumulateBug(new BugInstance(this, "REC2_CATCH_THROWABLE", LOW_PRIORITY).addClassAndMethod(this),
                             SourceLineAnnotation.fromVisitedInstruction(getClassContext(), this, caughtException.sourcePC));
                 }
             }
@@ -200,8 +216,20 @@ public class RuntimeExceptionCapture extends OpcodeStackDetector implements Stat
                             System.out.println("Dead exception store at " + first);
                         }
                         caughtException.dead = true;
-                        break;
                     }
+
+                    // See if the excpetion is rethrown at the end of the handler.
+                    while (block != null) {
+                        InstructionHandle last = block.getLastInstruction();
+                        if (last != null) {
+                            if (last.getInstruction().getOpcode() == Const.ATHROW) {
+                                caughtException.rethrown = true;
+                                break;
+                            }
+                        }
+                        block = cfg.getSuccessorWithEdgeType(block, EdgeTypes.FALL_THROUGH_EDGE);
+                    }
+                    break;
                 }
             }
         } catch (MethodUnprofitableException e) {